Frontrunner/etl/requirements.txt
mysql-connector-python==8.0.33
psycopg2-binary==2.9.9
pycryptodome==3.19.0

